% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress_tidy_groups.R
\name{suppress_tidy_data}
\alias{suppress_tidy_data}
\title{Suppress across multiple categories/groups within a single dataframe (with some additional options)}
\usage{
suppress_tidy_data(
  df,
  where_to_suppress = c("col"),
  cols_to_suppress,
  groups_to_suppress_across,
  suppression_pre_applied = FALSE,
  suppression_thres,
  suppress_0 = TRUE,
  inc_secondary_suppression = TRUE,
  suppression_output_value = 9999999,
  total_suppression = FALSE,
  secondary_suppress_0 = TRUE
)
}
\description{
This function allows users to suppress across multiple groups in a single data.
If you find yourself with a df that contains a column with multiple groups that you'd like to individually suppress on
you should use this function.
As an example, if you had a table with a region column (LONDON, MIDLANDS, etc) and you needed to suppress on each
region individually, you could use this function to shortcut the process.
}
